Removed code duplication
WaveformZoomView and WaveformOverview now inherit
a common base class, WaveformView
