+++ src/waveform-zoomview.js
@@ -6,17 +6,10 @@
  * @module waveform-zoomview
  */
 
-import PlayheadLayer from './playhead-layer';
-import PointsLayer from './points-layer';
-import SegmentsLayer from './segments-layer';
-import WaveformAxis from './waveform-axis';
-import WaveformShape from './waveform-shape';
+import WaveformView from './waveform-view';
 import InsertSegmentMouseDragHandler from './insert-segment-mouse-drag-handler';
 import ScrollMouseDragHandler from './scroll-mouse-drag-handler';
-import { clamp, formatTime, getMarkerObject, isFinite, isNumber, isValidTime,
-  objectHasProperty } from './utils';
-
-import Konva from 'konva/lib/Core';
+import { clamp, isValidTime, objectHasProperty } from './utils';
 
 /**
  * Creates a zoomable waveform view.
@@ -32,13 +25,7 @@ import Konva from 'konva/lib/Core';
 function WaveformZoomView(waveformData, container, peaks) {
   const self = this;
 
-  self._container = container;
-  self._peaks = peaks;
-  self._options = peaks.options;
-  self._viewOptions = self._options.zoomview;
-  self._enableWaveformCache = self._options.waveformCache;
-  self._originalWaveformData = waveformData;
-  self._initWaveformCache();
+  WaveformView.call(self, waveformData, container, peaks, peaks.options.zoomview);
 
   // Bind event handlers
   self._onTimeUpdate = self._onTimeUpdate.bind(self);
@@ -61,95 +48,40 @@ function WaveformZoomView(waveformData, container, peaks) {
   self._autoScroll = self._viewOptions.autoScroll;
   self._autoScrollOffset = self._viewOptions.autoScrollOffset;
 
-  self._amplitudeScale = 1.0;
-  self._timeLabelPrecision = self._viewOptions.timeLabelPrecision;
-  self._enableSeek = true;
   self._enableSegmentDragging = false;
   self._segmentDragMode = 'overlap';
   self._minSegmentDragWidth = 0;
   self._insertSegmentShape = null;
 
-  if (self._viewOptions.formatPlayheadTime) {
-    self._formatPlayheadTime = self._viewOptions.formatPlayheadTime;
-  }
-  else {
-    self._formatPlayheadTime = function(time) {
-      return formatTime(time, self._timeLabelPrecision);
-    };
-  }
-
   self._playheadClickTolerance = self._viewOptions.playheadClickTolerance;
-
-  self._data = null;
-  self._pixelLength = 0;
-
-  const initialZoomLevel = peaks.zoom.getZoomLevel();
 
   self._zoomLevelAuto = false;
   self._zoomLevelSeconds = null;
-
-  self._resampleData({ scale: initialZoomLevel });
-
-  self._width = container.clientWidth;
-  self._height = container.clientHeight;
-
-  // The pixel offset of the current frame being displayed
-  self._frameOffset = 0;
-
-  self._stage = new Konva.Stage({
-    container: container,
-    width: self._width,
-    height: self._height
-  });
-
-  self._waveformColor = self._viewOptions.waveformColor;
-  self._playedWaveformColor = self._viewOptions.playedWaveformColor;
-
-  self._createWaveform();
-
-  if (self._viewOptions.enableSegments) {
-    self._segmentsLayer = new SegmentsLayer(peaks, self, true);
-    self._segmentsLayer.addToStage(self._stage);
-  }
-
-  if (self._viewOptions.enablePoints) {
-    self._pointsLayer = new PointsLayer(peaks, self, true);
-    self._pointsLayer.addToStage(self._stage);
-  }
-
-  self._createAxisLabels();
-
-  self._playheadLayer = new PlayheadLayer({
-    player: self._peaks.player,
-    view: self,
-    showPlayheadTime: self._viewOptions.showPlayheadTime,
-    playheadColor: self._viewOptions.playheadColor,
-    playheadTextColor: self._viewOptions.playheadTextColor,
-    playheadFontFamily: self._viewOptions.fontFamily,
-    playheadFontSize: self._viewOptions.fontSize,
-    playheadFontStyle: self._viewOptions.fontStyle
-  });
-
-  self._playheadLayer.addToStage(self._stage);
 
   const time = self._peaks.player.getCurrentTime();
 
   self._syncPlayhead(time);
 
-  self._mouseDragHandler = new ScrollMouseDragHandler(self._peaks, self);
+  self._mouseDragHandler = new ScrollMouseDragHandler(peaks, self);
 
   self._onWheel = self._onWheel.bind(self);
   self._onWheelCaptureVerticalScroll = self._onWheelCaptureVerticalScroll.bind(self);
   self.setWheelMode(self._viewOptions.wheelMode);
 
-  self._onClick = self._onClick.bind(self);
-  self._onDblClick = self._onDblClick.bind(self);
-  self._onContextMenu = self._onContextMenu.bind(self);
-
-  self._stage.on('click', self._onClick);
-  self._stage.on('dblclick', self._onDblClick);
-  self._stage.on('contextmenu', self._onContextMenu);
+  self._peaks.emit('zoomview.displaying', 0, self.getEndTime());
 }
+
+WaveformZoomView.prototype = Object.create(WaveformView.prototype);
+
+WaveformZoomView.prototype.initWaveform = function() {
+  this._enableWaveformCache = this._options.waveformCache;
+
+  this._initWaveformCache();
+
+  const initialZoomLevel = this._peaks.zoom.getZoomLevel();
+
+  this._resampleData({ scale: initialZoomLevel });
+};
 
 WaveformZoomView.prototype._initWaveformCache = function() {
   if (this._enableWaveformCache) {
@@ -159,75 +91,7 @@ WaveformZoomView.prototype._initWaveformCache = function() {
   }
 };
 
-WaveformZoomView.prototype.enableSeek = function(enable) {
-  this._enableSeek = enable;
-};
-
-WaveformZoomView.prototype.isSeekEnabled = function() {
-  return this._enableSeek;
-};
-
-WaveformZoomView.prototype._onClick = function(event) {
-  this._clickHandler(event, 'click');
-};
-
-WaveformZoomView.prototype._onDblClick = function(event) {
-  this._clickHandler(event, 'dblclick');
-};
-
-WaveformZoomView.prototype._onContextMenu = function(event) {
-  this._clickHandler(event, 'contextmenu');
-};
-
-WaveformZoomView.prototype._clickHandler = function(event, eventName) {
-  let emitViewEvent = true;
-
-  if (event.target !== this._stage) {
-    const marker = getMarkerObject(event.target);
-
-    if (marker) {
-      if (marker.attrs.name === 'point-marker') {
-        const point = marker.getAttr('point');
-
-        if (point) {
-          this._peaks.emit('points.' + eventName, {
-            point: point,
-            evt: event.evt,
-            preventViewEvent: function() {
-              emitViewEvent = false;
-            }
-          });
-        }
-      }
-      else if (marker.attrs.name === 'segment-overlay') {
-        const segment = marker.getAttr('segment');
-
-        if (segment) {
-          const clickEvent = {
-            segment: segment,
-            evt: event.evt,
-            preventViewEvent: function() {
-              emitViewEvent = false;
-            }
-          };
-
-          if (this._segmentsLayer) {
-            this._segmentsLayer.segmentClicked(eventName, clickEvent);
-          }
-        }
-      }
-    }
-  }
-
-  if (emitViewEvent) {
-    const mousePosX = event.evt.layerX;
-    const time = this.pixelOffsetToTime(mousePosX);
-
-    this._peaks.emit('zoomview.' + eventName, {
-      time: time,
-      evt: event.evt
-    });
-  }
+WaveformZoomView.prototype.initHighlightLayer = function() {
 };
 
 WaveformZoomView.prototype.setWheelMode = function(mode, options) {
@@ -414,19 +278,6 @@ WaveformZoomView.prototype.getPlayheadClickTolerance = function() {
   return this._playheadClickTolerance;
 };
 
-WaveformZoomView.prototype.playheadPosChanged = function(time) {
-  if (this._playedWaveformShape) {
-    this._playedSegment.endTime = time;
-    this._unplayedSegment.startTime = time;
-
-    this._drawWaveformLayer();
-  }
-};
-
-WaveformZoomView.prototype.updatePlayheadTime = function(time) {
-  this._playheadLayer.updatePlayheadTime(time);
-};
-
 WaveformZoomView.prototype._syncPlayhead = function(time) {
   this._playheadLayer.updatePlayheadTime(time);
 
@@ -592,14 +443,6 @@ WaveformZoomView.prototype._resampleData = function(options) {
   this._pixelLength = this._data.length;
 };
 
-WaveformZoomView.prototype.getStartTime = function() {
-  return this.pixelOffsetToTime(0);
-};
-
-WaveformZoomView.prototype.getEndTime = function() {
-  return this.pixelOffsetToTime(this._width);
-};
-
 WaveformZoomView.prototype.setStartTime = function(time) {
   if (time < 0) {
     time = 0;
@@ -613,179 +456,11 @@ WaveformZoomView.prototype.setStartTime = function(time) {
 };
 
 /**
- * Returns the pixel index for a given time, for the current zoom level.
- *
- * @param {Number} time Time, in seconds.
- * @returns {Number} Pixel index.
- */
-
-WaveformZoomView.prototype.timeToPixels = function(time) {
-  return Math.floor(time * this._data.sample_rate / this._data.scale);
-};
-
-/**
- * Returns the time for a given pixel index, for the current zoom level.
- *
- * @param {Number} pixels Pixel index.
- * @returns {Number} Time, in seconds.
- */
-
-WaveformZoomView.prototype.pixelsToTime = function(pixels) {
-  return pixels * this._data.scale / this._data.sample_rate;
-};
-
-/**
- * Returns the time for a given pixel offset (relative to the
- * current scroll position), for the current zoom level.
- *
- * @param {Number} offset Offset from left-visible-edge of view
- * @returns {Number} Time, in seconds.
- */
-
-WaveformZoomView.prototype.pixelOffsetToTime = function(offset) {
-  const pixels = this._frameOffset + offset;
-
-  return pixels * this._data.scale / this._data.sample_rate;
-};
-
-WaveformZoomView.prototype.timeToPixelOffset = function(time) {
-  return Math.floor(time * this._data.sample_rate / this._data.scale) - this._frameOffset;
-};
-
-/**
- * @returns {Number} The start position of the waveform shown in the view,
- *   in pixels.
- */
-
-WaveformZoomView.prototype.getFrameOffset = function() {
-  return this._frameOffset;
-};
-
-/**
- * @returns {Number} The width of the view, in pixels.
- */
-
-WaveformZoomView.prototype.getWidth = function() {
-  return this._width;
-};
-
-/**
- * @returns {Number} The height of the view, in pixels.
- */
-
-WaveformZoomView.prototype.getHeight = function() {
-  return this._height;
-};
-
-/**
  * @returns {Number} The length of the waveform, in pixels.
  */
+
 WaveformZoomView.prototype.getPixelLength = function() {
   return this._pixelLength;
-};
-
-/**
- * @returns {Number} The media duration, in seconds.
- */
-
-WaveformZoomView.prototype._getDuration = function() {
-  return this._peaks.player.getDuration();
-};
-
-/**
- * Adjusts the amplitude scale of waveform shown in the view, which allows
- * users to zoom the waveform vertically.
- *
- * @param {Number} scale The new amplitude scale factor
- */
-
-WaveformZoomView.prototype.setAmplitudeScale = function(scale) {
-  if (!isNumber(scale) || !isFinite(scale)) {
-    throw new Error('view.setAmplitudeScale(): Scale must be a valid number');
-  }
-
-  this._amplitudeScale = scale;
-
-  this._drawWaveformLayer();
-
-  if (this._segmentsLayer) {
-    this._segmentsLayer.draw();
-  }
-};
-
-WaveformZoomView.prototype.getAmplitudeScale = function() {
-  return this._amplitudeScale;
-};
-
-/**
- * @returns {WaveformData} The view's waveform data.
- */
-
-WaveformZoomView.prototype.getWaveformData = function() {
-  return this._data;
-};
-
-WaveformZoomView.prototype._createWaveformShapes = function() {
-  if (!this._waveformShape) {
-    this._waveformShape = new WaveformShape({
-      color: this._waveformColor,
-      view: this
-    });
-
-    this._waveformShape.addToLayer(this._waveformLayer);
-  }
-
-  if (this._playedWaveformColor && !this._playedWaveformShape) {
-    const time = this._peaks.player.getCurrentTime();
-
-    this._playedSegment = {
-      startTime: 0,
-      endTime: time
-    };
-
-    this._unplayedSegment = {
-      startTime: time,
-      endTime: this._getDuration()
-    };
-
-    this._waveformShape.setSegment(this._unplayedSegment);
-
-    this._playedWaveformShape = new WaveformShape({
-      color: this._playedWaveformColor,
-      view: this,
-      segment: this._playedSegment
-    });
-
-    this._playedWaveformShape.addToLayer(this._waveformLayer);
-  }
-};
-
-WaveformZoomView.prototype._destroyPlayedWaveformShape = function() {
-  this._waveformShape.setSegment(null);
-
-  this._playedWaveformShape.destroy();
-  this._playedWaveformShape = null;
-
-  this._playedSegment = null;
-  this._unplayedSegment = null;
-};
-
-WaveformZoomView.prototype._createWaveform = function() {
-  this._waveformLayer = new Konva.Layer({ listening: false });
-
-  this._createWaveformShapes();
-
-  this._stage.add(this._waveformLayer);
-
-  this._peaks.emit('zoomview.displaying', 0, this.getEndTime());
-};
-
-WaveformZoomView.prototype._createAxisLabels = function() {
-  this._axisLayer = new Konva.Layer({ listening: false });
-  this._axis = new WaveformAxis(this, this._viewOptions);
-
-  this._axis.addToLayer(this._axisLayer);
-  this._stage.add(this._axisLayer);
 };
 
 /**
@@ -838,7 +513,7 @@ WaveformZoomView.prototype.updateWaveform = function(frameOffset) {
 
   this._playheadLayer.updatePlayheadTime(this.pixelsToTime(playheadPixel));
 
-  this._drawWaveformLayer();
+  this.drawWaveformLayer();
   this._axisLayer.draw();
 
   const frameStartTime = this.getStartTime();
@@ -855,65 +530,11 @@ WaveformZoomView.prototype.updateWaveform = function(frameOffset) {
   this._peaks.emit('zoomview.displaying', frameStartTime, frameEndTime);
 };
 
-WaveformZoomView.prototype._drawWaveformLayer = function() {
-  this._waveformLayer.draw();
-};
-
-WaveformZoomView.prototype.setWaveformColor = function(color) {
-  this._waveformColor = color;
-  this._waveformShape.setWaveformColor(color);
-};
-
-WaveformZoomView.prototype.setPlayedWaveformColor = function(color) {
-  this._playedWaveformColor = color;
-
-  if (color) {
-    if (!this._playedWaveformShape) {
-      this._createWaveformShapes();
-    }
-
-    this._playedWaveformShape.setWaveformColor(color);
-  }
-  else {
-    if (this._playedWaveformShape) {
-      this._destroyPlayedWaveformShape();
-    }
-  }
-};
-
-WaveformZoomView.prototype.showPlayheadTime = function(show) {
-  this._playheadLayer.showPlayheadTime(show);
-};
-
-WaveformZoomView.prototype.setTimeLabelPrecision = function(precision) {
-  this._timeLabelPrecision = precision;
-  this._playheadLayer.updatePlayheadText();
-};
-
-WaveformZoomView.prototype.formatTime = function(time) {
-  return this._formatPlayheadTime(time);
-};
-
-WaveformZoomView.prototype.showAxisLabels = function(show, options) {
-  this._axis.showAxisLabels(show, options);
-  this._axisLayer.draw();
-};
-
 WaveformZoomView.prototype.enableAutoScroll = function(enable, options) {
   this._autoScroll = enable;
 
   if (objectHasProperty(options, 'offset')) {
     this._autoScrollOffset = options.offset;
-  }
-};
-
-WaveformZoomView.prototype.enableMarkerEditing = function(enable) {
-  if (this._segmentsLayer) {
-    this._segmentsLayer.enableEditing(enable);
-  }
-
-  if (this._pointsLayer) {
-    this._pointsLayer.enableEditing(enable);
   }
 };
 
@@ -925,59 +546,36 @@ WaveformZoomView.prototype.setMinSegmentDragWidth = function(width) {
   this._minSegmentDragWidth = width;
 };
 
-WaveformZoomView.prototype.fitToContainer = function() {
-  if (this._container.clientWidth === 0 && this._container.clientHeight === 0) {
-    return;
-  }
-
+WaveformZoomView.prototype.containerWidthChange = function() {
   let updateWaveform = false;
 
-  if (this._container.clientWidth !== this._width) {
-    this._width = this._container.clientWidth;
-    this._stage.width(this._width);
+  let resample = false;
+  let resampleOptions;
 
-    let resample = false;
-    let resampleOptions;
+  if (this._zoomLevelAuto) {
+    resample = true;
+    resampleOptions = { width: this._width };
+  }
+  else if (this._zoomLevelSeconds !== null) {
+    resample = true;
+    resampleOptions = { scale: this._getScale(this._zoomLevelSeconds) };
+  }
 
-    if (this._zoomLevelAuto) {
-      resample = true;
-      resampleOptions = { width: this._width };
+  if (resample) {
+    try {
+      this._resampleData(resampleOptions);
+      updateWaveform = true;
     }
-    else if (this._zoomLevelSeconds !== null) {
-      resample = true;
-      resampleOptions = { scale: this._getScale(this._zoomLevelSeconds) };
-    }
-
-    if (resample) {
-      try {
-        this._resampleData(resampleOptions);
-        updateWaveform = true;
-      }
-      catch (error) {
-        // Ignore, and leave this._data as it was
-      }
+    catch (error) {
+      // Ignore, and leave this._data as it was
     }
   }
 
-  if (this._container.clientHeight !== this._height) {
-    this._height = this._container.clientHeight;
-    this._stage.height(this._height);
+  return updateWaveform;
+};
 
-    this._waveformShape.fitToView();
-    this._playheadLayer.fitToView();
-
-    if (this._segmentsLayer) {
-      this._segmentsLayer.fitToView();
-    }
-
-    if (this._pointsLayer) {
-      this._pointsLayer.fitToView();
-    }
-  }
-
-  if (updateWaveform) {
-    this.updateWaveform(this._frameOffset);
-  }
+WaveformZoomView.prototype.containerHeightChange = function() {
+  // Nothing
 };
 
 WaveformZoomView.prototype.getStage = function() {
@@ -988,15 +586,11 @@ WaveformZoomView.prototype.getSegmentsLayer = function() {
   return this._segmentsLayer;
 };
 
-WaveformZoomView.prototype.getViewOptions = function() {
-  return this._viewOptions;
-};
-
 WaveformZoomView.prototype.destroy = function() {
   // Unregister event handlers
-  this._peaks.off('player.timeupdate', this._onTimeUpdate);
   this._peaks.off('player.playing', this._onPlaying);
   this._peaks.off('player.pause', this._onPause);
+  this._peaks.off('player.timeupdate', this._onTimeUpdate);
   this._peaks.off('keyboard.left', this._onKeyboardLeft);
   this._peaks.off('keyboard.right', this._onKeyboardRight);
   this._peaks.off('keyboard.shift_left', this._onKeyboardShiftLeft);
@@ -1004,20 +598,7 @@ WaveformZoomView.prototype.destroy = function() {
 
   this._mouseDragHandler.destroy();
 
-  this._playheadLayer.destroy();
-
-  if (this._segmentsLayer) {
-    this._segmentsLayer.destroy();
-  }
-
-  if (this._pointsLayer) {
-    this._pointsLayer.destroy();
-  }
-
-  if (this._stage) {
-    this._stage.destroy();
-    this._stage = null;
-  }
+  WaveformView.prototype.destroy.call(this);
 };
 
 export default WaveformZoomView;
